# backend/openapi/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "ReprocessDocumentContentDto": {
        "properties": {},
        "type": "object"
      },
...
      "UpdateDocumentDto": {
        "properties": {},
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Document Management System API",
    "title": "DMS API",
    "version": "1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/documents": {
      "get": {
        "operationId": "DocumentsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "documents"
        ]
      }
    },
    "/documents/reprocess-content": {
      "post": {
        "operationId": "DocumentsController_reprocessContent",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReprocessDocumentContentDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reprocesa extracción de texto y OCR para documentos existentes",
        "tags": [
          "documents"
        ]
      }
    },
    "/documents/upload": {
      "post": {
        "operationId": "DocumentsController_upload",
        "parameters": [],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "areaCode": {
                    "example": "RC",
                    "type": "string"
                  },
                  "categoryId": {
                    "example": 1,
                    "type": "number"
                  },
                  "comentario": {
                    "example": "Primera version",
                    "type": "string"
                  },
                  "consecutivo": {
                    "example": 1,
                    "type": "number"
                  },
                  "documentTypeCode": {
                    "example": "PRO",
                    "type": "string"
                  },
                  "file": {
                    "format": "binary",
                    "type": "string"
                  },
                  "nombreDocumento": {
                    "example": "Procedimiento Seguridad",
                    "type": "string"
                  }
                },
                "required": [
                  "file",
                  "nombreDocumento"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "documents"
        ]
      }
    },
    "/documents/{id}": {
      "get": {
        "operationId": "DocumentsController_findOne",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "versionsLimit",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "documents"
        ]
      },
      "patch": {
        "operationId": "DocumentsController_update",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDocumentDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "documents"
        ]
      }
    },
    "/documents/{id}/approve": {
      "post": {
        "operationId": "DocumentsController_approveDecision",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "comentario": {
                    "example": "OK",
                    "type": "string"
                  },
                  "decision": {
                    "example": "APPROVED",
                    "type": "string"
                  }
                },
                "required": [
                  "decision"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Approver decision",
        "tags": [
          "documents"
        ]
      }
    },
    "/documents/{id}/assign-reviewers": {
      "patch": {
        "operationId": "DocumentsController_assignReviewers",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "aproboUserId": {
                    "example": 3,
                    "type": "number"
                  },
                  "revisoUserId": {
                    "example": 2,
                    "type": "number"
                  }
                },
                "required": [
                  "revisoUserId",
                  "aproboUserId"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Assign reviewer and approver",
        "tags": [
          "documents"
        ]
      }
    },
    "/documents/{id}/obsolete": {
      "post": {
        "operationId": "DocumentsController_markObsolete",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark document as obsolete",
        "tags": [
          "documents"
        ]
      }
    },
    "/documents/{id}/review": {
      "post": {
        "operationId": "DocumentsController_reviewDecision",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "comentario": {
                    "example": "OK",
                    "type": "string"
                  },
                  "decision": {
                    "example": "APPROVED",
                    "type": "string"
                  }
                },
                "required": [
                  "decision"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reviewer decision",
        "tags": [
          "documents"
        ]
      }
    },
    "/documents/{id}/submit-review": {
      "post": {
        "operationId": "DocumentsController_submitReview",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Submit document for review",
        "tags": [
          "documents"
        ]
      }
    },
    "/documents/{id}/versions": {
      "get": {
        "operationId": "DocumentsController_listVersions",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "documents"
        ]
      }
    },
    "/documents/{id}/workflow": {
      "get": {
        "operationId": "DocumentsController_getWorkflow",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "documents"
        ]
      }
    },
...
  },
  "servers": [],
...
}